requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.59.6
mammoth==1.8.0
pypdf==4.3.1
openpyxl==3.1.5
httpx==0.27.2
pytest==8.3.4
pytest-cov==5.0.0
pytest-playwright==0.5.2
playwright==1.50.0
